requirements.txt
# Core dependencies
python-docx==1.1.0
openai==1.12.0
anthropic==0.18.1
deepl==1.18.0
requests==2.31.0
tiktoken>=0.8.0
python-dotenv==1.0.0

